api_web_scrapping/requirements.txt
requests
BeautifulSoup4
uvicorn
fastapi
pandas
tabula-py